+++ shared/rollouts/features.py
@@ -2,4 +2,4 @@ from . import Feature
 
 BUNDLE_THRESHOLD_FLAG = Feature("bundle_threshold_flag")
 INCLUDE_GITHUB_COMMENT_ACTIONS_BY_OWNER = Feature("include_github_comment_actions")
-USE_NEW_MINIO = Feature("use_new_minio")
+NEW_MINIO = Feature("new_minio")
