+++ shared/storage/__init__.py
@@ -1,39 +1,30 @@
-from typing import Literal
+from typing import Literal, cast
 
 from shared.config import get_config
-from shared.rollouts.features import USE_NEW_MINIO
-from shared.storage.aws import AWSStorageService
+from shared.rollouts.features import NEW_MINIO
 from shared.storage.base import BaseStorageService
-from shared.storage.fallback import StorageWithFallbackService
-from shared.storage.gcp import GCPStorageService
 from shared.storage.minio import MinioStorageService
-from shared.storage.new_minio import NewMinioStorageService
-
-_storage_service_cache: dict[str, BaseStorageService] = {}
 
 
 def get_appropriate_storage_service(
     repoid: int | None = None,
     force_minio=False,
 ) -> BaseStorageService:
-    chosen_storage = "minio"
-    if repoid and USE_NEW_MINIO.check_value(repoid, default=False):
-        chosen_storage = "new_minio"
+    return get_minio_storage_service(repoid)
 
-    if chosen_storage not in _storage_service_cache:
-        _storage_service_cache[chosen_storage] = (
-            _get_appropriate_storage_service_given_storage(chosen_storage)
+
+def get_minio_storage_service(
+    repo_id: int | None,
+) -> MinioStorageService:
+    minio_config = get_config("services", "minio", default={})
+    if repo_id:
+        new_minio_mode = cast(
+            Literal["read", "write"] | None,
+            NEW_MINIO.check_value(repo_id, default=None),  # type: ignore
         )
-
-    return _storage_service_cache[chosen_storage]
-
-
-def _get_appropriate_storage_service_given_storage(
-    chosen_storage: Literal["minio", "new_minio"],
-) -> BaseStorageService:
-    if chosen_storage == "new_minio":
-        minio_config = get_config("services", "minio", default={})
-        return NewMinioStorageService(minio_config)
-    elif chosen_storage == "minio":
-        minio_config = get_config("services", "minio", default={})
+        return MinioStorageService(
+            minio_config,
+            new_mode=new_minio_mode,
+        )
+    else:
         return MinioStorageService(minio_config)
